+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexParameterImpl.java
@@ -64,4 +64,34 @@ public class ImpexParameterImpl extends ImpexParameterMixin implements ImpexPara
     return findChildByClass(ImpexSubParameters.class);
   }
 
+  @Override
+  @Nullable
+  public String getReferenceItemTypeName() {
+    return ImpexPsiUtil.getReferenceItemTypeName(this);
+  }
+
+  @Override
+  @Nullable
+  public String getReferenceName() {
+    return ImpexPsiUtil.getReferenceName(this);
+  }
+
+  @Override
+  @Nullable
+  public String getItemTypeName() {
+    return ImpexPsiUtil.getItemTypeName(this);
+  }
+
+  @Override
+  @Nullable
+  public String getInlineTypeName() {
+    return ImpexPsiUtil.getInlineTypeName(this);
+  }
+
+  @Override
+  @NotNull
+  public String getAttributeName() {
+    return ImpexPsiUtil.getAttributeName(this);
+  }
+
 }