+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexUnknownTypeAttributeInspection.kt
@@ -23,52 +23,60 @@ import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
-import com.intellij.idea.plugin.hybris.impex.psi.*
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexAnyHeaderParameterName
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexParameter
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.DOCUMENT_ID
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexVisitor
+import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexFunctionTSAttributeReference
+import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexTSAttributeReference
 import com.intellij.idea.plugin.hybris.psi.reference.TSReferenceBase
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.psi.impl.source.tree.LeafPsiElement
-import com.intellij.psi.util.PsiTreeUtil
 
-/**
- * @author Nosov Aleksandr <nosovae.dev@gmail.com>
- */
 class ImpexUnknownTypeAttributeInspection : LocalInspectionTool() {
+
     override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.ERROR
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor = ImpexHeaderLineVisitor(holder)
 
     private class ImpexHeaderLineVisitor(private val problemsHolder: ProblemsHolder) : ImpexVisitor() {
 
         override fun visitParameter(parameter: ImpexParameter) {
-            inspect(parameter)
+            if (parameter.firstChild is ImpexMacroUsageDec || (parameter.firstChild as? LeafPsiElement)?.elementType == DOCUMENT_ID) return
+
+            parameter.references
+                .find { it is ImpexFunctionTSAttributeReference }
+                ?.let { it as? ImpexFunctionTSAttributeReference }
+                ?.takeIf { it.multiResolve(false).isEmpty() }
+                ?.let { ref ->
+                    parameter.itemTypeName
+                        ?.let { registerProblem(ref, it) }
+                }
+                ?: return
         }
 
         override fun visitAnyHeaderParameterName(parameter: ImpexAnyHeaderParameterName) {
-            inspect(parameter)
-        }
-
-        private fun inspect(parameter: PsiElement) {
             if (parameter.firstChild is ImpexMacroUsageDec || (parameter.firstChild as? LeafPsiElement)?.elementType == DOCUMENT_ID) return
 
-            val firstReference = parameter.references.firstOrNull() ?: return
-
-            if (firstReference.canonicalText.contains(".")) return
-            if (firstReference !is TSReferenceBase<*>) return
-            if (firstReference.multiResolve(false).isNotEmpty()) return
-
-            val typeName = (
-                PsiTreeUtil.getParentOfType(parameter, ImpexParameter::class.java)
-                    ?: PsiTreeUtil.getParentOfType(parameter, ImpexFullHeaderParameter::class.java)
-                        ?.anyHeaderParameterName
-                )
-                ?.text
-                ?: "?"
-            problemsHolder.registerProblem(
-                parameter,
-                message("hybris.inspections.UnknownTypeAttributeInspection.key", parameter.text, typeName),
-                ProblemHighlightType.ERROR)
+            parameter.references
+                .find { it is ImpexTSAttributeReference }
+                ?.let { it as? ImpexTSAttributeReference }
+                ?.takeIf { it.multiResolve(false).isEmpty() }
+                ?.let { ref ->
+                    parameter.headerItemTypeName
+                        ?.text
+                        ?.let { registerProblem(ref, it) }
+                }
+                ?: return
         }
 
+        private fun registerProblem(reference: TSReferenceBase<out PsiElement>, typeName: String) {
+            problemsHolder.registerProblemForReference(
+                reference,
+                ProblemHighlightType.ERROR,
+                message("hybris.inspections.UnknownTypeAttributeInspection.key", reference.value, typeName),
+            )
+        }
     }
 }