+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/completion/impl/DefaultTSCompletionService.kt
@@ -36,41 +36,13 @@ class DefaultTSCompletionService(private val project: Project) : TSCompletionSer
 
     override fun getCompletions(typeCode: String) = getCompletions(
         typeCode,
-        TSMetaType.META_ITEM, TSMetaType.META_ENUM, TSMetaType.META_RELATION
+        TSMetaType.META_ITEM, TSMetaType.META_ENUM, TSMetaType.META_RELATION, TSMetaType.META_COLLECTION, TSMetaType.META_MAP
     )
 
     override fun getCompletions(typeCode: String, vararg types: TSMetaType) = getCompletions(
         typeCode,
         0, *types
     )
-
-    fun getCompletions(typeCode: String, recursionLevel: Int, vararg types: TSMetaType): List<LookupElementBuilder> {
-        if (recursionLevel > HybrisConstants.TS_MAX_RECURSION_LEVEL) return emptyList()
-
-        val metaService = TSMetaModelAccess.getInstance(project)
-        return types
-            .firstNotNullOfOrNull { metaType ->
-                when (metaType) {
-                    TSMetaType.META_ITEM -> metaService.findMetaItemByName(typeCode)
-                        ?.let { getCompletions(it) }
-
-                    TSMetaType.META_ENUM -> metaService.findMetaEnumByName(typeCode)
-                        ?.let { getCompletionsForEnum(metaService) }
-
-                    TSMetaType.META_RELATION -> metaService.findMetaRelationByName(typeCode)
-                        ?.let { getCompletions(it, metaService) }
-
-                    TSMetaType.META_COLLECTION -> metaService.findMetaCollectionByName(typeCode)
-                        ?.let { getCompletions(it.elementType, recursionLevel + 1, *types) }
-
-                    TSMetaType.META_MAP -> metaService.findMetaMapByName(typeCode)
-                        ?.let { getCompletions(it) }
-
-                    else -> null
-                }
-            }
-            ?: emptyList()
-    }
 
     override fun getCompletions(vararg types: TSMetaType) = with(TSMetaModelAccess.getInstance(project)) {
         types
@@ -100,6 +72,34 @@ class DefaultTSCompletionService(private val project: Project) : TSCompletionSer
                 }
             }
             .flatten()
+    }
+
+    private fun getCompletions(typeCode: String, recursionLevel: Int, vararg types: TSMetaType): List<LookupElementBuilder> {
+        if (recursionLevel > HybrisConstants.TS_MAX_RECURSION_LEVEL) return emptyList()
+
+        val metaService = TSMetaModelAccess.getInstance(project)
+        return types
+            .firstNotNullOfOrNull { metaType ->
+                when (metaType) {
+                    TSMetaType.META_ITEM -> metaService.findMetaItemByName(typeCode)
+                        ?.let { getCompletions(it) }
+
+                    TSMetaType.META_ENUM -> metaService.findMetaEnumByName(typeCode)
+                        ?.let { getCompletionsForEnum(metaService) }
+
+                    TSMetaType.META_RELATION -> metaService.findMetaRelationByName(typeCode)
+                        ?.let { getCompletions(it, metaService) }
+
+                    TSMetaType.META_COLLECTION -> metaService.findMetaCollectionByName(typeCode)
+                        ?.let { getCompletions(it.elementType, recursionLevel + 1, *types) }
+
+                    TSMetaType.META_MAP -> metaService.findMetaMapByName(typeCode)
+                        ?.let { getCompletions(it) }
+
+                    else -> null
+                }
+            }
+            ?: emptyList()
     }
 
     private fun getCompletionsForEnum(metaModelAccess: TSMetaModelAccess) = metaModelAccess.findMetaItemByName(HybrisConstants.TS_TYPE_ENUMERATION_VALUE)